requirements-test.txt
# Testing requirements for unified admin system
# These are additional requirements for comprehensive testing

# Core testing framework
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0
pytest-xdist>=3.3.1
pytest-mock>=3.11.1
pytest-html>=3.2.0
pytest-json-report>=1.5.0

# End-to-end testing
playwright>=1.37.0
pytest-playwright>=0.4.2

# Performance testing
pytest-benchmark>=4.0.0
locust>=2.16.1

# Security testing
bandit>=1.7.5
safety>=2.3.5

# Code quality
flake8>=6.0.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.1

# Coverage reporting
coverage>=7.3.0
coverage-badge>=1.1.0

# Database testing
factory-boy>=3.3.0
faker>=19.3.0

# API testing
requests>=2.31.0
responses>=0.23.3

# Mocking and fixtures
freezegun>=1.2.2
time-machine>=2.12.0

# Django testing utilities
django-test-plus>=2.2.0
django-webtest>=1.9.11
model-bakery>=1.12.0

# Multi-tenant testing
django-tenants>=3.5.0

# Authentication testing
django-otp>=1.2.2

# Persian/RTL testing utilities
python-bidi>=0.4.2
arabic-reshaper>=3.0.0

# Reporting and documentation
sphinx>=7.1.2
sphinx-rtd-theme>=1.3.0

# Development utilities
ipdb>=0.13.13
django-debug-toolbar>=4.2.0
django-extensions>=3.2.3

# Environment management
python-dotenv>=1.0.0

# Async testing
pytest-asyncio>=0.21.1

# Browser automation
selenium>=4.11.2

# Image comparison for UI tests
pillow>=10.0.0
opencv-python>=4.8.0.76

# Network testing
httpretty>=1.1.4

# Time zone testing
pytz>=2023.3

# JSON schema validation
jsonschema>=4.19.0

# Load testing
artillery>=1.7.9

# Memory profiling
memory-profiler>=0.61.0
psutil>=5.9.5

# Logging testing
structlog>=23.1.0

# Internationalization testing
babel>=2.12.1

# File handling testing
openpyxl>=3.1.2
xlsxwriter>=3.1.2

# Email testing
django-mail-panel>=4.0.2

# Cache testing
django-redis>=5.3.0

# Task queue testing
celery>=5.3.1
kombu>=5.3.1

# Monitoring testing
prometheus-client>=0.17.1

# Health check testing
django-health-check>=3.17.0